requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
drf-yasg
python-dotenv
black
requests
